Separate the two implementations of TDP. This also allows for a future third implementation (BearSSL?).
